PYTHON/requirements.txt
# DATA DIFF PYTHON - Dependencies
# ================================
# Using datacompy with native Snowflake/Snowpark integration

# Core comparison library with Snowflake support
datacompy[snowflake]>=0.11.1

# Snowpark (included with datacompy[snowflake], but listed explicitly)
snowflake-snowpark-python>=1.11.0

# Data manipulation
pandas>=2.0.0
numpy>=1.24.0

# Configuration management
python-dotenv>=1.0.0

# Excel export support
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Progress bars and CLI
tqdm>=4.66.0
click>=8.1.0
rich>=13.7.0

# Logging and reporting
tabulate>=0.9.0

# YAML config support
pyyaml>=6.0.0
